+++ src/components/layouts/Tab/TabList.tsx
@@ -1,13 +1,7 @@
 import styled from '@emotion/styled';
 import { Justify } from '../../../models/FlexTypes';
-import {
-	Children,
-	PropsWithChildren,
-	useLayoutEffect,
-	useRef,
-	useState,
-} from 'react';
-import { MIN_TAB_LENGTH } from '../../../constants';
+import { PropsWithChildren } from 'react';
+import useTabButton from './hooks/useTabButton';
 
 export interface TableListProps extends PropsWithChildren {
 	justify?: Justify;
@@ -18,37 +12,8 @@ export interface TableListProps extends PropsWithChildren {
 const TabList = (props: TableListProps) => {
 	const { justify = 'left', onClick, moveButton = false, children } = props;
 
-	const tabContainerRef = useRef<HTMLDivElement>(null);
-	const tabListRef = useRef<HTMLDivElement>(null);
-
-	const [showButton, setShowButton] = useState(false);
-	const countArray = Children.toArray(children).length;
-
-	useLayoutEffect(() => {
-		if (!tabContainerRef.current) return;
-		const { width } = tabContainerRef.current.getBoundingClientRect();
-		const isOverFlow = moveButton && width <= countArray * MIN_TAB_LENGTH;
-
-		setShowButton(isOverFlow);
-	}, [countArray, moveButton]);
-
-	const scrollTabHandler = (event: React.MouseEvent<HTMLButtonElement>) => {
-		const tabList = tabListRef.current;
-		const containerRect = tabContainerRef.current?.getBoundingClientRect();
-
-		if (!showButton || !tabList || !containerRect) return;
-		const { dataset } = event.currentTarget;
-
-		if (dataset['direction'] === 'prev') {
-			tabList.scrollLeft -= containerRect.width;
-			return;
-		}
-
-		if (dataset['direction'] === 'next') {
-			tabList.scrollLeft += containerRect.width;
-			return;
-		}
-	};
+	const { tabContainerRef, scrollTabHandler, showButton, tabListRef } =
+		useTabButton({ children, moveButton });
 
 	return (
 		<Wrapper ref={tabContainerRef} onClick={onClick} justify={justify}>
