+++ src/components/layouts/Tab/hooks/useTabButton.tsx
@@ -1,0 +1,45 @@
+import { Children, useLayoutEffect, useRef, useState } from 'react';
+import { MIN_TAB_LENGTH } from '../../../../constants';
+
+interface useTabButtonProps {
+	children: React.ReactNode;
+	moveButton: boolean;
+}
+
+const useTabButton = ({ children, moveButton }: useTabButtonProps) => {
+	const tabContainerRef = useRef<HTMLDivElement>(null);
+	const tabListRef = useRef<HTMLDivElement>(null);
+
+	const [showButton, setShowButton] = useState(false);
+	const countArray = Children.toArray(children).length;
+
+	useLayoutEffect(() => {
+		if (!tabContainerRef.current) return;
+		const { width } = tabContainerRef.current.getBoundingClientRect();
+		const isOverFlow = moveButton && width <= countArray * MIN_TAB_LENGTH;
+
+		setShowButton(isOverFlow);
+	}, [countArray, moveButton]);
+
+	const scrollTabHandler = (event: React.MouseEvent<HTMLButtonElement>) => {
+		const tabList = tabListRef.current;
+		const containerRect = tabContainerRef.current?.getBoundingClientRect();
+
+		if (!showButton || !tabList || !containerRect) return;
+		const { dataset } = event.currentTarget;
+
+		if (dataset['direction'] === 'prev') {
+			tabList.scrollLeft -= containerRect.width;
+			return;
+		}
+
+		if (dataset['direction'] === 'next') {
+			tabList.scrollLeft += containerRect.width;
+			return;
+		}
+	};
+
+	return { tabContainerRef, tabListRef, scrollTabHandler, showButton };
+};
+
+export default useTabButton;
